Fixed minor URL parsing mistakes
In the case that a requested URL begins with "//", we replace this with the current protocol that we're operating under i.e. "http://", "https://", etc.

In the case that a requested URL begins with more than two or a single "/", we simply concatenate it to the current page's "scheme://fqdn"